requirement.txt
alembic==1.4.1
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autograd==1.3
autograd-gamma==0.5.0
azure-core==1.9.0
azure-storage-blob==12.6.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
catboost==0.24.3
certifi==2020.6.20
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorlover==0.3.0
combo==0.1.2
confuse==1.4.0
cryptography==3.3.1
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
databricks-cli==0.14.1
decorator==4.4.2
DeepPurpose==0.0.7
defusedxml==0.6.0
descriptastorus==2.0.0.32
docker==4.4.1
entrypoints==0.3
Flask==1.1.2
funcy==1.15
future==0.18.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
graphviz==0.16
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
imbalanced-learn==0.7.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1607970656719/work
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver @ file:///D:/bld/kiwisolver_1604322452227/work
kmodes==0.10.2
lifelines==0.25.7
lightgbm==3.1.1
llvmlite==0.35.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib @ file:///D:/bld/matplotlib-suite_1605180506412/work
missingno==0.4.2
mistune==0.8.4
mkl-service==2.3.0
mlflow==1.12.1
mlxtend==0.18.0
msrest==0.6.19
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.1.6
numba==0.52.0
numexpr==2.7.1
numpy @ file:///D:/bld/numpy_1604946157939/work
oauthlib==3.1.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging==20.8
pandas @ file:///D:/bld/pandas_1607423802728/work
pandas-flavor==0.2.0
pandas-profiling==2.9.0
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
phik==0.10.0
pickleshare==0.7.5
Pillow @ file:///D:/bld/pillow_1604748918199/work
plac==1.1.3
plotly==4.14.1
pluggy==0.13.1
preshed==3.0.5
prettytable==2.0.0
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.8
protobuf==3.14.0
py==1.10.0
pyarrow==2.0.0
pycairo==1.20.0
pycaret==2.2.3
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.3
pyLDAvis==2.1.2
pyod==0.8.5
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1604321279890/work
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
querystring-parser==1.2.4
regex==2020.11.13
reportlab==3.5.56
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn @ file:///C:/ci/scikit-learn_1598377018496/work
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
smart-open==4.0.1
smmap==3.0.4
spacy==2.3.5
SQLAlchemy @ file:///D:/bld/sqlalchemy_1608367002681/work
sqlparse==0.4.1
srsly==1.0.5
statsmodels==0.12.1
streamlit==0.73.1
subword-nmt==0.3.7
suod==0.0.4
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
thinc==7.4.5
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
toolz==0.11.1
torch==1.7.1
tornado @ file:///D:/bld/tornado_1604105353952/work
tqdm==4.54.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
umap-learn==0.4.6
urllib3==1.26.2
validators==0.18.2
visions==0.5.0
waitress==1.4.4
wasabi==0.8.0
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.1
xarray==0.16.2
xgboost==1.3.1
yellowbrick==1.2
